TX and RX working, need to verify timing. Small updates to report
usart.c/dma.c
- killed the rx dma channel, using IT mode for rx

MPU9250.c
- added a small comment

freertos.c
- added new objects (queues, queuesets etc)
- added tx and rx task
- added basic inter-task comm mechanisms for tx
- commented out the gyro data acquisition since apparently we aren't using it

CANRGX-Embedded.tex
- updated packet description
- added short calculation for upper byte rate limit

+++ MPUTest/Inc/MPU9250.h
@@ -109,7 +109,7 @@ typedef struct{
 	float az; // Acceleration along z-axis
 	float ay; // Acceleration along y-axis
 	float ax; // Acceleration along x-axis
-	double A;
+	float A;  // ||a||
 	float vz; // Yaw rate (about z-axis)
 	float vy; // Pitch rate (about y-axis)
 	float vx; // Roll rate (about x-axis)
